% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Data import function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the location of the data file.}
}
\value{
a tibble of FARS data for a given year
}
\description{
This function is used to take data from the US-NHTSA Fatality Analysis Reporting System (FARS)
in a .csv file and return it back as a tidyverse dataframe(tibble).
}
\note{
Function will also check is file exists and throw an error if it is not present.
}
\examples{
fars_read("accident_2014.csv.bz2")

}
