% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS summary data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector of calendar years}
}
\value{
a tibble of incidents by month per year
}
\description{
This function will process the FARS data to produce a tibble that shows the number of
fatalities by month per year
}
\examples{
fars_summarize_years(2013:2015)
}
