% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year as either character or numeric string}
}
\value{
character string
}
\description{
Function to make a custom filename based on a given year. Given an input year
(ex. \code{1983}) it will return a character string (\code{[1] "accident_1983.csv.bz2"})
}
\examples{
\dontrun{
make_filename(1983)
}
}
